new class file for userfeedbacklink
